% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/km_plot_weighted.R
\name{km_grouped_weighted}
\alias{km_grouped_weighted}
\title{Kaplan Meier plots for two groups with IPTW}
\usage{
km_grouped_weighted(
  data,
  time,
  event,
  group,
  weight_col,
  group_name,
  endpoint,
  title = "",
  x_title = "Time",
  y_title = "Adjusted Survival Probability",
  time_survival = 1,
  risk_table = TRUE,
  x_lim = NULL,
  y_lim = c(0, 1),
  x_breaks = waiver(),
  y_breaks = seq(0, 1, by = 0.2),
  custom_colors = NULL,
  text_size_title = 15,
  text_size = 12,
  show_label = FALSE,
  unit = "months",
  color_curves = FALSE,
  show_weighted_n = TRUE,
  show_p_values = TRUE,
  p_placement = c(0.05, 0.15),
  legend_placement = c(0.6, 0.25)
)
}
\arguments{
\item{data}{A `data.frame` containing the necessary columns: time, event, and group.}

\item{time}{A character string specifying the column name of the numeric time
variable for the time-to-event endpoint.}

\item{event}{A character string specifying the column name of the event indicator.
Should be a numeric variable where 1 represents an event, and 0
represents no event or right-censoring.}

\item{group}{A character string specifying the grouping variable used for stratifying
the Kaplan-Meier curves.}

\item{weight_col}{Character string indicating the column name containing the weights.}

\item{group_name}{Names of the groups to be compared (used for the risk table).}

\item{endpoint}{A character string providing the name of the endpoint.}

\item{title}{A character string providing the title for the Kaplan-Meier plot.}

\item{x_title}{A character string specifying the label for the x-axis.}

\item{y_title}{A character string specifying the label for the y-axis.}

\item{time_survival}{A numeric value representing the time point at which the
event pobability is calculated. Default is 1.}

\item{risk_table}{Logical, default: TRUE. Should a risk table be displayed?}

\item{x_lim}{A numeric vector of length two defining the limits of the x-axis.}

\item{y_lim}{A numeric vector of length two defining the limits of the y-axis.
Default is `c(0, 1)`.}

\item{x_breaks}{A numeric vector specifying the break points for the x-axis labels.}

\item{y_breaks}{A numeric vector specifying the break points for the y-axis labels.
Default is from 0 to 1 with steps of 0.2.}

\item{custom_colors}{Character string specifying the colors for the risk table. If not provided, default colors will be automatically selected.}

\item{text_size_title}{A numeric value specifying the font size of the title.}

\item{text_size}{A numeric value specifying the font size of the text elements.}

\item{show_label}{A logical or character value indicating whether to display median survival time of event probabilites
labels within the plot. Options are `FALSE` (default), `median`, or `probability`.}

\item{unit}{A character string indicating the unit for `time_survival` (e.g., "months").
Default is "months".}

\item{color_curves}{Character string specifying the colors for the survival curves.}

\item{show_weighted_n}{Logical, default: TRUE. Should the weighted or unweighted number of observations at risk be displayed?}

\item{show_p_values}{logical, indicating whether the p-values be displayed in the plot?. Default = `TRUE`}

\item{p_placement}{numeric vector of length 2 indicating the position of the the p-value.
The first number correspondend to the hjust placement, the second to the vjust placement}

\item{legend_placement}{numeric vector of length 2 indicating the legend position. Default is c(0.6, 0.2).}
}
\value{
# list containing
a, the ggplot object of the Kaplan-Meier plot
b, a data frame containing median survival, the probability to get an event after
the prespecified period and its confidence interval, the p value of the log rank or wilcoxon test
}
\description{
This function plots Kaplan-Meier curves and calculates median survival time,
the probability of getting an event after a certain period and p values of
either the log-rank test or Wilcoxon test for two strata
}
