% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cum_incidence_grouped.R
\name{comp_risk_grouped}
\alias{comp_risk_grouped}
\title{Function for cumulative incidence plots of two groups with competing events}
\usage{
comp_risk_grouped(
  data,
  time,
  event,
  group,
  title = "",
  x_title = waiver(),
  y_title = waiver(),
  time_survival = 1,
  unit = "months",
  endpoint = "",
  x_lim = NULL,
  y_lim = c(0, 1),
  x_breaks = waiver(),
  y_breaks = seq(0, 1, by = 0.2),
  colors = FALSE,
  show_label = "none",
  text_size_title = 15,
  text_size = 12,
  p_placement = c(0.05, 0.15),
  legend_placement = c(0.6, 0.8),
  time_vec_prob = c(1, 2, 3)
)
}
\arguments{
\item{data}{A `data.frame` containing the necessary columns: time, event, and group.}

\item{time}{A character string specifying the column name of the numeric time
variable for the time-to-event endpoint. 0 = censored.}

\item{event}{A character string specifying the column name of the event indicator.
Should be a numeric variable where 1 represents an event, and 0
represents no event or right-censoring.}

\item{group}{A character string specifying the grouping variable used for stratifying
the Kaplan-Meier curves.}

\item{title}{A character string providing the title for the Kaplan-Meier plot.}

\item{x_title}{A character string specifying the label for the x-axis.}

\item{y_title}{A character string specifying the label for the y-axis.}

\item{time_survival}{A numeric value representing the time point at which the
event pobability is calculated. Default is 1.}

\item{unit}{A character string indicating the unit for `time_survival` (e.g., "months").
Default is "months".}

\item{endpoint}{A character string providing the name of the endpoint.}

\item{x_lim}{A numeric vector of length two defining the limits of the x-axis.}

\item{y_lim}{A numeric vector of length two defining the limits of the y-axis.
Default is `c(0, 1)`.}

\item{x_breaks}{A numeric vector specifying the break points for the x-axis labels.}

\item{y_breaks}{A numeric vector specifying the break points for the y-axis labels.
Default is from 0 to 1 with steps of 0.2.}

\item{colors}{A character vector specifying the colors for the curves. If not provided,
random colors will be assigned.}

\item{show_label}{A logical or character value indicating whether to display median survival time of event probabilites
labels within the plot. Options are `FALSE` (default), `median`, or `probability`.}

\item{text_size_title}{A numeric value specifying the font size of the title.}

\item{text_size}{A numeric value specifying the font size of the text elements.}

\item{p_placement}{numeric vector of length 2 indicating the position of the the p-value.
The first number correspondend to the hjust placement, the second to the vjust placement}

\item{legend_placement}{numeric vector of length 2 indicating the legend position. Default is c(0.6, 0.2).}

\item{time_vec_prob}{numeric vector containing time points, on which one wants to display
the probability of the incidence.}
}
\value{
a list with the cumulative incidence plot, a table with the median incidence
and a table with the probability of an event after certain timepoints
}
\description{
This function plots the cumualative incidence curves of two groups with competing
events. Plottet will only be the first argument. One can optinally display the
median incidence time or incidence after a certain time.
}
